Remove debug print statements.
