+++ libs/ledger-lib-persistence/src/main/kotlin/net/corda/ledger/libs/persistence/util/NamedParamStatement.kt
@@ -1,6 +1,9 @@
 package net.corda.ledger.libs.persistence.util
 
 import net.corda.crypto.core.InvalidParamsException
+import net.corda.ledger.libs.persistence.utxo.impl.UtxoPersistenceServiceImpl
+import org.slf4j.Logger
+import org.slf4j.LoggerFactory
 import java.sql.Connection
 import java.sql.PreparedStatement
 import java.sql.ResultSet
@@ -9,59 +12,13 @@ import java.time.Instant
 import java.util.Calendar
 import java.util.TimeZone
 
-interface NamedParamQuery {
-    companion object {
-        fun from(sql: String): NamedParamQuery {
-            val plainSql = StringBuilder()
-            val fields = mutableMapOf<String, Int>()
-            var marker: StringBuilder? = null
-            var markerIndex = 0
-            for (i in sql.indices) {
-                val c = sql[i]
-                if (c == ':' && (i < sql.length - 1 && sql[i + 1].isValidTokenChar())) {
-                    marker = StringBuilder()
-                    markerIndex++
-                    plainSql.append('?')
-                    continue
-                }
-
-                if (null != marker) {
-                    if (!c.isValidTokenChar()) {
-                        fields[marker.toString()] = markerIndex
-                        marker = null
-                    } else {
-                        marker.append(c)
-                        continue
-                    }
-                }
-                plainSql.append(c)
-            }
-
-            if (null != marker) {
-                fields[marker.toString()] = markerIndex
-            }
-
-            return NamedParamQueryImpl(plainSql.toString(), fields)
-        }
-
-        private fun Char.isValidTokenChar(): Boolean = this.isLetterOrDigit() || this == '_' || this == '-'
-    }
-
-    val sql: String
-    val fields: Map<String, Int>
-
-    private class NamedParamQueryImpl(
-        override val sql: String,
-        override val fields: Map<String, Int>,
-    ) : NamedParamQuery
-}
-
 class NamedParamStatement(
     private val namedParamQuery: NamedParamQuery,
     private val conn: Connection,
 ) : AutoCloseable {
-    companion object {
+    private companion object {
         val tzUTC: Calendar = Calendar.getInstance(TimeZone.getTimeZone("UTC"))
+        val log: Logger = LoggerFactory.getLogger(UtxoPersistenceServiceImpl::class.java)
     }
     private val statement: PreparedStatement = conn.prepareStatement(namedParamQuery.sql)
 
@@ -70,7 +27,7 @@ class NamedParamStatement(
     }
 
     fun executeQuery(): ResultSet {
-        println("############ $statement")
+        if(log.isDebugEnabled) log.debug("Execute Query: $statement")
         return statement.executeQuery()
     }
 
@@ -100,7 +57,8 @@ class NamedParamStatement(
     }
 
     fun executeUpdate(): Int {
-        println("############ $statement")
+        if(log.isDebugEnabled) log.debug("Execute Update: $statement")
+
         return statement.executeUpdate()
     }
 
@@ -113,7 +71,6 @@ class NamedParamStatement(
 
     fun setStrings(name: String, values: List<String>) {
         val stringsArray = conn.createArrayOf("varchar", values.toTypedArray())
-        println("#### filtering: $values")
         statement.setArray(
             getFieldIndex(name),
             stringsArray
